Push everything into a separate namespace
