+++ bag_replicator/src/main/scala/uk/ac/wellcome/platform/archive/bagreplicator/services/BagStorageService.scala
@@ -1,7 +1,6 @@
 package uk.ac.wellcome.platform.archive.bagreplicator.services
 
 import grizzled.slf4j.Logging
-import uk.ac.wellcome.platform.archive.bagreplicator.config.ReplicatorDestinationConfig
 import uk.ac.wellcome.platform.archive.common.models.bagit.BagLocation
 import uk.ac.wellcome.storage.s3.S3PrefixCopier
