+++ bag_replicator_unpack_to_archive/src/main/scala/uk/ac/wellcome/platform/archive/bagreplicator_unpack_to_archive/Main.scala
@@ -1,14 +1,11 @@
-package uk.ac.wellcome.platform.archive.bagreplicator
+package uk.ac.wellcome.platform.archive.bagreplicator_unpack_to_archive
 
 import akka.actor.ActorSystem
 import com.typesafe.config.Config
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.typesafe.{NotificationStreamBuilder, SNSBuilder}
-import uk.ac.wellcome.platform.archive.bagreplicator.config.BagReplicatorConfig
-import uk.ac.wellcome.platform.archive.bagreplicator.services.{
-  BagReplicatorWorkerService,
-  BagStorageService
-}
+import uk.ac.wellcome.platform.archive.bagreplicator_unpack_to_archive.config.BagReplicatorConfig
+import uk.ac.wellcome.platform.archive.bagreplicator_unpack_to_archive.services.{BagReplicatorWorkerService, BagStorageService, UnpackedBagService}
 import uk.ac.wellcome.platform.archive.common.models.BagRequest
 import uk.ac.wellcome.storage.s3.S3PrefixCopier
 import uk.ac.wellcome.storage.typesafe.S3Builder
@@ -24,6 +21,8 @@ object Main extends WellcomeTypesafeApp {
     implicit val executionContext: ExecutionContextExecutor =
       actorSystem.dispatcher
 
+    val s3Client = S3Builder.buildS3Client(config)
+
     val s3PrefixCopier = S3PrefixCopier(
       s3Client = S3Builder.buildS3Client(config)
     )
@@ -36,6 +35,8 @@ object Main extends WellcomeTypesafeApp {
       notificationStream =
         NotificationStreamBuilder.buildStream[BagRequest](config),
       bagStorageService = bagStorageService,
+      unpackedBagService = new UnpackedBagService(s3Client),
+      s3PrefixCopier = S3PrefixCopier(s3Client),
       bagReplicatorConfig = BagReplicatorConfig.buildBagReplicatorConfig(config),
       progressSnsWriter =
         SNSBuilder.buildSNSWriter(config, namespace = "progress"),
