+++ bag_replicator_unpack_to_archive/src/main/scala/uk/ac/wellcome/platform/archive/bagreplicator_unpack_to_archive/config/ReplicatorDestinationConfig.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.bagreplicator.config
+package uk.ac.wellcome.platform.archive.bagreplicator_unpack_to_archive.config
 
 // Specifies the S3 bucket and root path for objects copied by the replicator.
 case class ReplicatorDestinationConfig(namespace: String,
