+++ bag_replicator_unpack_to_archive/src/test/scala/uk/ac/wellcome/platform/archive/bagreplicator_unpack_to_archive/services/BagReplicatorWorkerServiceTest.scala
@@ -1,18 +1,12 @@
-package uk.ac.wellcome.platform.archive.bagreplicator.services
+package uk.ac.wellcome.platform.archive.bagreplicator_unpack_to_archive.services
 
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.json.JsonUtil._
-import uk.ac.wellcome.platform.archive.bagreplicator.fixtures.{
-  BagReplicatorFixtures,
-  WorkerServiceFixture
-}
+import uk.ac.wellcome.platform.archive.bagreplicator_unpack_to_archive.fixtures.{BagReplicatorFixtures, WorkerServiceFixture}
 import uk.ac.wellcome.platform.archive.common.fixtures.BagLocationFixtures
 import uk.ac.wellcome.platform.archive.common.generators.BagRequestGenerators
-import uk.ac.wellcome.platform.archive.common.models.bagit.{
-  BagLocation,
-  BagPath
-}
+import uk.ac.wellcome.platform.archive.common.models.bagit.{BagLocation, BagPath}
 import uk.ac.wellcome.platform.archive.common.models.ReplicationResult
 import uk.ac.wellcome.platform.archive.common.progress.ProgressUpdateAssertions
 import uk.ac.wellcome.platform.archive.common.progress.models.Progress
